Api Fixes| Get Enrollment new api added

+++ src/main/java/com/lti/mod/services/adminservice/service/AdminService.java
@@ -12,7 +12,7 @@ public interface AdminService {
 	Object createTechnology(Technology technology);
 	void deleteCourse(Long parseLong);
 	Object createUser(User user);
-	void deleteUser(BigInteger id);
+	void deleteUser(Long id);
 	List<User> findAllUsers();
 	List<User> findAllMentors();
 }
